pyproject.toml
[project]
name = "al-anime-recommender-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Farhan", email = "farhanrhine@gmail.com" }
]
dependencies = [
    "chromadb>=1.3.5",
    "fastapi>=0.123.8",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.0",
    "langchain-huggingface>=1.1.0",
    "pandas>=2.3.3",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "slowapi>=0.1.9",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
]

[project.urls]
Homepage = "https://github.com/farhanrhine/Al-Anime-Recommender-System-using-Grafana-Cloud-Minikube-ChromaDB-Langchain-GCP"
Source = "https://github.com/farhanrhine/Al-Anime-Recommender-System-using-Grafana-Cloud-Minikube-ChromaDB-Langchain-GCP.git"



